add utf8 io tests

+++ t/103-thruk_utils_io.t
@@ -1,6 +1,5 @@
 use warnings;
 use strict;
-use Encode qw/encode_utf8/;
 use File::Temp qw/tempfile/;
 use Test::More;
 use utf8;
@@ -8,7 +7,7 @@ use utf8;
 use Thruk::Utils::Encode ();
 
 plan skip_all => 'internal test only' if defined $ENV{'PLACK_TEST_EXTERNALSERVER_URI'};
-plan tests => 14;
+plan tests => 53;
 
 use_ok('Thruk::Utils::IO');
 use_ok('Thruk::Config');
@@ -18,6 +17,9 @@ my $testdata = { data => [
     12345,
     { nested => ['list'] },
     '€',
+    'öäüß',
+    '的',
+    '€ öäüß 的',
 ]};
 my($fh, $file) = tempfile();
 unlink($file);
@@ -32,6 +34,17 @@ is_deeply($testdata, $readdata, 'data is the same');
 
 ok(unlink($file), 'remove tempfile');
 
+################################################################################
+for my $data (@{$testdata->{'data'}}) {
+    Thruk::Utils::IO::json_lock_store($file, [$data]);
+    ok(-f $file, "file does not exist now: ".$file);
+
+    my $readdata = Thruk::Utils::IO::json_lock_retrieve($file);
+    is_deeply([$data], $readdata, 'data is the same');
+    ok(unlink($file), 'remove tempfile');
+}
+
+################################################################################
 my $hostname = `hostname`;
 my(undef, $hostname2) = Thruk::Utils::IO::cmd("hostname");
 is($hostname2, $hostname, "hostnames are equal");
@@ -41,21 +54,17 @@ is($hostname3, $hostname, "hostnames are equal");
 
 ################################################################################
 # write utf8
-my $teststr = 'öäüß';
-my $rc = Thruk::Utils::IO::write($file, $teststr);
-is($rc, 1, "write succeeded");
+for my $teststr (@{$testdata->{'data'}}) {
+    next if ref $teststr;
+    use Encode;
+    my $rc = Thruk::Utils::IO::write($file, $teststr);
+    is($rc, 1, "write succeeded");
 
-my $str = Thruk::Utils::IO::read($file);
-is($str, $teststr, "read string matched");
-ok(unlink($file), 'remove tempfile');
+    my $str = Thruk::Utils::IO::read($file);
+    $str = Thruk::Utils::Encode::decode_any($str);
+    is($str, $teststr, "read string matched");
 
-################################################################################
-# write decoded utf8
-$teststr = Thruk::Utils::Encode::decode_any($teststr);
-$rc = Thruk::Utils::IO::write($file, $teststr);
-is($rc, 1, "write succeeded");
-
-$str = Thruk::Utils::IO::read($file);
-is($str, $teststr, "read string matched");
-
-ok(unlink($file), 'remove tempfile');
+    (undef, $str) = Thruk::Utils::IO::cmd("cat ".$file);
+    is($str, $teststr, "read string from cat matched");
+    ok(unlink($file), 'remove tempfile');
+}
